+++ angular_compiler/lib/src/outliner.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/constant/value.dart';
 import 'package:build/build.dart';
+import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 
 import 'analyzer.dart';
@@ -40,7 +41,7 @@ class TemplateOutliner implements Builder {
 
   TemplateOutliner(
     this._compilerFlags, {
-    String extension: '.outline.template.dart',
+    @required String extension,
   })
       : _extension = extension,
         buildExtensions = {
@@ -62,7 +63,7 @@ class TemplateOutliner implements Builder {
     final injectors = <String>[];
     var units = [library.definingCompilationUnit]..addAll(library.parts);
     var types = units.expand((unit) => unit.types);
-    var methods = units.expand((unit) => unit.functions);
+    var fields = units.expand((unit) => unit.topLevelVariables);
     for (final clazz in types) {
       final component = $Component.firstAnnotationOfExact(
         clazz,
@@ -80,12 +81,12 @@ class TemplateOutliner implements Builder {
         }
       }
     }
-    for (final method in methods) {
+    for (final field in fields) {
       if ($GenerateInjector.hasAnnotationOfExact(
-        method,
+        field,
         throwOnUnresolved: false,
       )) {
-        injectors.add('${method.name}\$Injector');
+        injectors.add('${field.name}\$Injector');
       }
     }
     final output = new StringBuffer('$_analyzerIgnores\n');